Safari-609.1.11.1/LayoutTests/fast/scrolling/scroll-to-focused-element-asynchronously-expected.txt

This tests that scrolling to the focused element happens asynchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.scrollingElement.scrollTop is 0
PASS document.scrollingElement.scrollTop > 100 is true
PASS successfullyParsed is true

TEST COMPLETE
target
